[consensus] prepare for separate safety rules process
This refactors the code so that it can conveniently be leveraged in the
process model. Specifically the common code across thread-approach.

Closes: #2208
Approved by: davidiw
