+++ consensus/safety-rules/Cargo.toml
@@ -12,6 +12,7 @@ consensus-types = { path = "../consensus-types", version = "0.1.0" }
 lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-config = { path = "../../config", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
+libra-tools = { path = "../../common/tools", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 serde = { version = "1.0.99", default-features = false }
 thiserror = "1.0"
